Add workaround for TargetVersion issue on VS 2017
Sometimes when installing for Visual Studio the version of
`ServiceHub.VSDetouredHost.exe` will be used instead of `devenv.exe`.
This means that the version of Visual Studio 2017 can appears as 1.x
instead of the expected 15.x.

This workaround installs the `GitClientVS.TeamFoundation.15` MEF asset
when the `TargetVersion` is 1.x (as well as the expected 15.x).